move enums to remove circular dependency
